requirements.txt
colorama
ipwhois==1.1.0
validators==0.14.2
argcomplete==1.11.1
dnspython==1.16.0
